fix(console): UTC-pin date formatting + move formatters to lib/format
Addresses the #330 review:
- Blocking (SDE2): formatDate rendered ISO dates in local time, so viewers west
  of UTC saw the prior day ("May 27" for 2026-05-28). Pin the date formatter to
  timeZone: 'UTC' so the authored calendar day renders everywhere. (The CI/dev
  browser runs in a UTC+ zone, which is why the browser-walk didn't catch it.)
- Minor (Architect): move the generic formatCurrency/formatDate to lib/format.ts
  (the plan's home for formatters; Billing/Analytics will reuse them) and leave
  the CRM-specific ContactStatusBadge/initials in contact-ui.tsx.
