+++ apps/console/src/modules/crm/contact-ui.tsx
@@ -24,17 +24,3 @@ export function initials(name: string): string {
   const last = parts.length > 1 ? (parts[parts.length - 1]?.[0] ?? '') : '';
   return (first + last).toUpperCase();
 }
-
-const currencyFmt = new Intl.NumberFormat('en-US', {
-  style: 'currency',
-  currency: 'USD',
-  maximumFractionDigits: 0,
-});
-const dateFmt = new Intl.DateTimeFormat('en-US', {
-  month: 'short',
-  day: 'numeric',
-  year: 'numeric',
-});
-
-export const formatCurrency = (value: number) => currencyFmt.format(value);
-export const formatDate = (iso: string) => dateFmt.format(new Date(iso));
